Create dir of right ownership (#82)
* Create dir of right ownership

See this comment in the commit:

OK, so the v.MountPath will _exist_ on the system, as systemd will create it, BUT the permissions/ownership might be wrong
We will chown the directory to the user/group of the security context, but the directory is created by systemd, when it
performs the automount. We'll add ExecPreStart that chowns the directory, but only if it's empty, otherwise we might be
messing with some system directory - this may still be the case when the dir is empty though. Because a StartExecPre to check
this would be a shell script we make the check here:
* create/chown if not exist
* chown if empty dir
* anything else don't touch: this likely makes the unit start, but then fail. (we may intercept this and return a decent message
  so you can see this with events

Cleanup of these directories is hard - systemd clean <unit> exists for doesn't assume other units are using this space.

Signed-off-by: Miek Gieben <miek@miek.nl>

* More logging on the vks side

log what we're doing to those poor directories.

Signed-off-by: Miek Gieben <miek@miek.nl>