Add provider testing (#88)
This works by having podspec yaml and concatentated units file in the
testdir. We call CreatePod and see what unit files have been delivered
to systemd for execution. This is just string diffing in the final
stage, but very useful to see what we should generate.

Needed some changes in the createpod code to make failing a directory a
non-error. A better approach is probably a DryRun boolean in provider
that will not touch disk when tue.

See: #76

This adds unit (hah!) testing.

Signed-off-by: Miek Gieben <miek@miek.nl>