export node's IP addresses as env vars (#23)
* export node's IP addresses as env vars

This uncovered a heap of stuff that wasn't done properly, hence the size
of the PR.

This adds 2 new flags --node-ip and --node-external-ip, which override
whatever we find ourselves. We also set this on node object. And we
export these as SYSTEMK_NODE_INTERNAL_IP and SYSTEMK_NODE_EXTERNAL_IP
(unclear if we keep this though, looks handy though). Next I want to
test this setup with Prometheus rewriting targets based on these values.
The step after that is using CoreDNS which binds the metrics to the
internal IP address by using the SYSTEMK_NODE_INTERNAL_IP env var.

We also fix the environment settings for the API server host and port.
We do require a kubeconfig to be given and should probably fatal if not
seen.

Signed-off-by: Miek Gieben <miek@miek.nl>

* Fix prometheus pod yaml

Signed-off-by: Miek Gieben <miek@miek.nl>