Add startup probes
This PR adds a field in the provider that hold the status of the latest probe check.
When looking a pod's yaml we check the startup probe and initiate the
correct (either tcpSocket, or httpGet) probe with the correct values.

In the statsToContainer code we just retrieve the check's status and use
that in the container; this still needs to be wired up.

Signed-off-by: Miek Gieben <miek@miek.nl>