Add --dir -d flag to force mounts points be below a path
This adds extra security to force mountspoints to be below a specified
path list (defaults to "/var"), as to prevent system modifications.

This is also showed HostPaths don't work, but appeared to work because
everything was done as root.

Also remove stray println in tests and cleanup the test output (disable
logging)

Signed-off-by: Miek Gieben <miek@miek.nl>