Remove CGO dependency, but add journalctl binary one
This PR makes it possible to drop CGO and get a std static binary for
Go. It does so by removing the sdjournal dep, and instead launch the
journalctl process.

This PR ballooned a bit, because several flags were dropped or not
working, so it brings back:

* setting node name to `hostname` if nothing is specified, this only
  uses  the default of 'systemk' if `os.Hostname` fails somehow
* adds backs `--disable-taint` to remove the systemk taint that get's
  set.

Tailing logs works and doesn't leave a zombie journalctl process. But I
haven't implemented all functionality yet. It looks like `Previous`
can't be done at all using this route.

Signed-off-by: Miek Gieben <miek@miek.nl>