Remove user mode
This removes the -u flag and assorted code, i.e. we will always use the
system's systemd (and run as root).

Left a TODO that I will shortly address to map to a non-root user (if
that flag is set) if the unit file says run as root and no security
context is provided in the prodspec.

Signed-off-by: Miek Gieben <miek@miek.nl>