Fix nodename setting
The nodename is not working, the instalFlags would overwrite the
Opts with the default value in the flag string. So it would overwrite
it (in the case of nodename) with the empty value. This shows when
you don't have the HOSTNAME env var set (the case in my zsh setup)

Move this into the correct order, first set the flags, then set the
default options if the flag isn't set.

Signed-off-by: Miek Gieben <miek@miek.nl>