Add --internal|--external-iface
This takes the 1 address of the *interface* and uses this for internal
or external node IP. This saves a whole bunch of external parsing (ip
addr show | grep ...) to get this IP.

E.g. I've got another (more stable) address configured on lo:1, so just
specifying the iface would make the unit file more portible and less
fragile because no iface parsing scripts needs to be executed.

Signed-off-by: Miek Gieben <miek@miek.nl>