Support projected ServiceAccountTokens, ConfigMaps and Secrets
ServiceAccountToken are moved to a new volume API - projected volumes -
add the initial ground work to support those so that pod start up
doesn't fail. The only thing I don't understand yet is the expiration
and how that's supposed to work as the backing volume is still a secret,
should probably be added to the 'watching' list as well.

Refactor the path creation into setupPaths because that we all identical
code.

Also add support for projected secrets and configmaps.

Signed-off-by: Miek Gieben <miek@miek.nl>