Add e2e testing
Add a workflow that kicks of the e2e tests on a PR.

Add a test for check the output from one of the containers and add the
supporting code to make that possible.

Fix potential crash in systemk when a section is not found in a unit
because the unit is already being deleted from the system.

Export the journal stuff from the provider so we can use it in testutil.
Make a small wrapper for getting a single line from the logs.

Signed-off-by: Miek Gieben <miek@miek.nl>